\name{dat0}
\alias{dat0}
\title{Aggregated and anonymized dataset from Yahoo representing server metrics of Yahoo services}
\description{
  Aggregated and anonymized dataset from Yahoo representing server metrics of Yahoo services.
}

\usage{
  data(dat0)
}

\format{time series}

\author{
  Rob J Hyndman, Earo Wang, Nikolay Laptev
}

\keyword{datasets}
