\name{dat4}
\alias{dat4}
\title{Synthetic time-series data-set where 3 time-series have unusual amount of noise}
\description{
  Synthetic time-series data-set where 3 time-series have unusual amount of noise
}

\usage{
  data(dat4)
}

\format{time series}

\author{
  Rob J Hyndman, Earo Wang, Nikolay Laptev
}

\keyword{datasets}
