\name{dat5}
\alias{dat5}
\title{Synthetic time series }
\description{
  Synthetic time series. \code{dat5} contains 100 time series with 3 time series having an unusually large trend. Each time series consists of 1400 observations.
}

\examples{
  plot(dat5, plot.type='single')
  plot(dat5[,1:10])
}

\format{time series}

\author{
  Rob J Hyndman, Earo Wang, Nikolay Laptev
}

\keyword{datasets}
