\name{dat4}
\alias{dat4}
\title{Synthetic time series }
\description{
  Synthetic time series. \code{dat4} contains 100 time series with 3 time series having an unusually large amount of noise. Each time series consists of 1400 observations.
}

\examples{
  plot(dat4, plot.type='single')
  plot(dat4[,1:10])
}

\format{time series}

\author{
  Rob J Hyndman, Earo Wang, Nikolay Laptev
}

\keyword{datasets}
